refactor: implement dynamic bridge loading with resource detection
This change moves framework to a bridge level. Using file loads to only load the resource that is detected as started.

I would like this to be the structure used going forwards.

This eliminates the need for loading files the lib will never need to use for a particular setup, eliminates nested if/else statements, and the need to specifically define the resources to be used in the config.

TODO:
- Implement this same structure for the rest of the 'bridge' modules: inventory, emotes, etc.
- Remove ox_lib dependency for lib.load